Made a parallel version of code
I wrote a parallel version of the laplace solver. I had to re-write the maximum_change function to get this to work. The issue is that the max_change variable was private (the way the old function was written). This meant that there was no way to run a reduction clause, or to find out the maximum local change. To fix this, I made the Maximum_Chnage function pass a third argument, a pointer to the max_change variable (which is set as shared in the main parallel region). Everything is in a single parallel region. All timing data uses Omp_get_wtime (rather than clock()). Further, all timing data is performed by the master thread.