Initial commit. Rewrote Mesh variables for weak scaling
I rewrote the code to perform weak scaling. This was difficult to do with a non-square number of procs. I added a Is_Square function to check if the number of procs is a square number. If it is, the program continues. Otherwise it exits.

I also had to redo the way that the mesh variables are established. This is because the number of mesh points changes with the number of procs. C doesn't let you run function calls in global (static duration) variables. Thus, I had to move the definitions for N_Mesh, N_Rows, N_Cols, and change_threshold into the main function. To allow sub functions to work with these variables, I added appropiate paramaters to each function. The paramaters are passed as constants to ensure no accidential modification occurs. I also had to define dx, dy locally in both the BC and update functions.

I made appropiate changes to the shared/private variables lists for the parallel region.

Finally, I made changes to the printout section. The program now prints out the total number of meshpoints in each direction, and the the number of elements per thread.