Initial commit. Optimized Update function
In my origional program, the update function worked as follows:

First move U_kp1 to U_k. Then use U_k to update U_kp1.

The issue here is that we're always updating two matricies worth of elements. Both U_k and U_kp1 are modified by this process, but they don't need to be. This is what my optimization aims to fix. This optimization works by passing the iteration number in addition to the two matricies. The two  matricies have been renamed to U_Even and U_Odd. If the iteration numebr is even, then we use U_Odd to make U_Even. In other words, if the numebr of iterations is odd then U_k is stored in U_Even and U_kp1 will be stored in U_Odd. The flip is true for if the iteration number is odd. By doing this, we don't need to copy the elements over. Instead, we just need to figure out which matrix (even or odd) is storing U_k, and then use that to store U_kpe in the other matrix.

This means that in each step only one of the matricies are updated. This improves runtime. It should be noted, however, that this does not double runtime. This is because moving the matricies over involves far fewer operations than updating the elements.

The iteration update had to be placed in a single declaration. Without this, threads would start updating beforre the iteration number was set. This meant that some threads would think U_odd was U_k, while others thought U_Even was U_k. This made the results non-determinisitic. Placing Iter++ in a single (with implied barrier) fixed this problem.