Updated stopping condition
I changed the stopping condition from a linear relationship with N_Mesh to a quadratic one. I imperically found that a constant of .5 in the numerator worked.

To test if my analysis was indeed correct, I ran the old code on stampede with 1000 meshpoints. I then downloaded the results file and graphed it. Indeed, the solution did not properly converged/it dropped a lot in the middle. I then repeated this process with he updated stopping condition. With the quadratic stopping condition the resulting plot looked correct. Therefore, I believe that my analysis is correct!